+++ react/FixedButton.js
@@ -1,0 +1,3 @@
+import FixedButton from './components/FixedButton'
+
+export default FixedButton